[nfc] cast inputs to ctype.h 'is*' functions
The ctype.h `is*` functions are typically implemented as macros using a lookup
table. This is fine, but on Cygwin the compiler is somewhat picky about this:

  lib/ast/fmtesc.c:116:59: warning: array subscript has type char
    [-Wchar-subscripts]
    116 |                           (c == '#' && (b == f || isspace(*(b - 1)))
        |                                                           ^~~~~~~~

This change squashes these warnings.

+++ cmd/tools/graphml2gv.c
@@ -203,7 +203,7 @@ static int isAnonGraph(char *name)
 {
     if (*name++ != '%')
 	return 0;
-    while (isdigit(*name))
+    while (isdigit((int)*name))
 	name++;			/* skip over digits */
     return (*name == '\0');
 }
